Revert useless changes
